add trims
