+++ lib/stitches/__init__.py
@@ -1,4 +1,6 @@
 from running_stitch import *
 from auto_fill import auto_fill
-from auto_satin import auto_satin
 from fill import legacy_fill
+
+# Can't put this here because we get a circular import :(
+#from auto_satin import auto_satin
